pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lease-review-tool"
version = "0.1.0"
description = "Backend scaffold for a lease review knowledge-base and analysis pipeline."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "boto3>=1.35,<2.0",
  "fastapi>=0.115,<1.0",
  "uvicorn>=0.30,<1.0",
  "sqlalchemy>=2.0,<3.0",
  "psycopg[binary]>=3.2,<4.0",
  "pgvector>=0.3,<1.0",
  "pydantic>=2.7,<3.0",
  "pymupdf>=1.24,<2.0",
  "python-dotenv>=1.0,<2.0",
  "openai>=1.50.0,<2.0",
  "pypdf>=5.0,<6.0",
  "python-docx>=1.1,<2.0",
]

[project.scripts]
lease-review-tool = "lease_review_tool.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
